Remove the use of `.to_f` in Spree.
Floating point numbers are not suitable for use in representing money
due to a loss of precision.

There may still be areas where floating point numbers are used
internally, however with this change we are not willingly subjecting
ourselves to the horror of floating point numbers.

+++ core/app/models/spree/calculator/shipping/flat_percent_item_total.rb
@@ -15,7 +15,7 @@ module Spree
 
       def compute_from_price(price)
         value = price * BigDecimal(self.preferred_flat_percent.to_s) / 100.0
-        (value * 100).round.to_f / 100
+        (value * 100).round.to_d / 100
       end
     end
   end
