+++ core/app/models/spree/payment/processing.rb
@@ -39,7 +39,7 @@ module Spree
             gateway_options
           )
           money = ::Money.new(amount, currency)
-          capture_events.create!(amount: money.to_f)
+          capture_events.create!(amount: money.to_d)
           split_uncaptured_amount
           handle_response(response, :complete, :failure)
         end
