+++ core/spec/models/spree/calculator/default_tax_spec.rb
@@ -69,7 +69,7 @@ describe Spree::Calculator::DefaultTax, :type => :model do
           # ex pre-tax = $57.14
           # 57.14 + %5 = 59.997 (or "close enough" to $60)
           # 60 - 57.14 = $2.86
-          expect(calculator.compute(order).to_f).to eql 2.86
+          expect(calculator.compute(order).to_d).to eql 2.86
         end
       end
     end
