+++ core/spec/models/spree/order_contents_spec.rb
@@ -45,13 +45,13 @@ describe Spree::OrderContents, :type => :model do
     end
 
     it "should update order totals" do
-      expect(order.item_total.to_f).to eq(0.00)
-      expect(order.total.to_f).to eq(0.00)
+      expect(order.item_total.to_d).to eq(0.00)
+      expect(order.total.to_d).to eq(0.00)
 
       subject.add(variant, 1)
 
-      expect(order.item_total.to_f).to eq(19.99)
-      expect(order.total.to_f).to eq(19.99)
+      expect(order.item_total.to_d).to eq(19.99)
+      expect(order.total.to_d).to eq(19.99)
     end
 
     context "running promotions" do
@@ -138,17 +138,17 @@ describe Spree::OrderContents, :type => :model do
     end
 
     it "should update order totals" do
-      expect(order.item_total.to_f).to eq(0.00)
-      expect(order.total.to_f).to eq(0.00)
+      expect(order.item_total.to_d).to eq(0.00)
+      expect(order.total.to_d).to eq(0.00)
 
       subject.add(variant,2)
 
-      expect(order.item_total.to_f).to eq(39.98)
-      expect(order.total.to_f).to eq(39.98)
+      expect(order.item_total.to_d).to eq(39.98)
+      expect(order.total.to_d).to eq(39.98)
 
       subject.remove(variant,1)
-      expect(order.item_total.to_f).to eq(19.99)
-      expect(order.total.to_f).to eq(19.99)
+      expect(order.item_total.to_d).to eq(19.99)
+      expect(order.total.to_d).to eq(19.99)
     end
   end
